requirements.txt
request
random
string
time 
os
colorama
Fore
sys